ml-service 2/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
prophet>=1.1.4
scikit-learn>=1.3.0
lightfm>=1.17
networkx>=3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
supabase>=2.0.0
scipy>=1.11.0
transformers>=4.35.0
torch>=2.0.0
bertopic>=0.15.0
sentence-transformers>=2.2.0
umap-learn>=0.5.4
hdbscan>=0.8.33
shap>=0.43.0
lime>=0.2.0
tensorflow>=2.13.0
pytest>=7.4.0
